Add files via upload
The Anomaly Detection layer played an integral role in the core of our multi-layered 
analysis process. This layer laid the groundwork for our investigation into the 
fascinating world of TikTok in the context of the Turkish elections of 2023. The 
application of machine learning techniques, specifically the Isolation Forest algorithm, 
was central to our strategy, and the algorithm's effectiveness in detecting anomalies 
helped shed light on the complexities of the TikTok landscape during this politicallycharged period.
9
There were two reasons for incorporating the Isolation Forest algorithm. Our primary 
objective was to maximize our computational resources. By focusing our analysis on 
the most influential videos, we were able to significantly reduce the high 
computational cost associated with the application of facial and emotional recognition 
techniques in subsequent stages of the analysis. The second objective was to develop 
a comprehensive understanding of the engagement metrics on the TikTok platform, 
which we hoped to accomplish by identifying the 'impactful' videos or anomalies in 
our dataset.