Add files via upload
In the second layer of this multi-layered machine learning analysis, face detection and 
emotion recognition technologies are utilized. This phase's primary objective is to 
identify the emotions expressed by content creators affiliated with the two polarized 
political parties in question. Rather than emphasizing the recognition of specific 
politicians or party-related figures, the primary focus is on understanding the emotive 
messages conveyed by the content, irrespective of the individuals featured.
11
Video content associated to the political party CHP provides an illustration of this 
strategy. A video may not feature a prominent politician, but instead feature a satisfied 
citizen praising the CHP's initiatives. The happiness or contentment depicted in this 
video would then be interpreted as representative of the party's emotional tone, despite 
the absence of recognizable political figures.