Add files via upload
The third layer of our multi-layered TikTok political network analysis 
integrates speech recognition to identify the specific narratives and topics conveyed 
in the videos. By carefully analyzing the spoken content, we can gain a deeper 
understanding of the nuances of the ongoing political discussions. This layer is 
essential for the creation of additional features for each video, thereby enhancing our 
knowledge of the political leanings of the TikTok network.
The underlying assumption is that the subjects discussed by a content creator and the 
sentiments they express may significantly influence the prospective voters in the 
TikTok network. Deciphering the spoken content is therefore 
essential for determining the messages being transmitted, the topics of the discussions, 
and the nature of the emotions evoked.

We've designed two major functions for this layer of analysis that work in tandem. 
The initial section is dependent on audio extraction from the videos. For this purpose, 
we utilize MoviePy, a robust Python library for video editing that offers a variety of 
features, including audio extraction from video files. The primary 
class used for audio extraction is VideoFileClip, which provides access to the audio 
component of a video file.