Add files via upload
The fourth level of the multi-layered approach of our machine learning model focuses 
on applying natural language processing (NLP) to TikTok video transcripts and 
associated comments. This layer delves into linguistic 
analysis in order to comprehend the underlying topical structures and sentiments 
expressed in the video content and associated discussions. This step is crucial to our 
research because it provides us with nuanced insights into the semantic composition 
of the data and enables us to understand user sentiment towards various topics, both 
of which can be utilized to better comprehend the sociopolitical dynamics of the 
platform.