Add files via upload
Feature engineering bridges the gap between raw data and effective models. It involves creating 
new features from existing ones in order to improve the learning process and 
performance of machine learning models. In this TikTok video analysis project, 
feature engineering is essential to the generation of insights that could help us better 
comprehend the strategic maneuvers of two political parties. This layer of analysis is 
comprehensive, sophisticated, and indispensable for enhancing the preceding machine 
learning analysis layers.

Below you can find what does the features signal for, 

Engagement Rate: This is the first metric calculated, and it represents the ratio of likes 
to views. It measures the video's popularity or appeal. High engagement rates 
21
indicate not only passive consumption, but also active participation from users with 
the latter taking the additional step of approving the content as opposed to simply 
viewing it.
The 'Face_Detection_Rate_FE' metric measures the rate at which faces are detected in 
a video. These two characteristics reveal the audience-resonating strategies employed 
by the parties. A high face detection rate may indicate that parties are employing 
recognizable faces to gain support.
Face Detection Rate: This metric represents the frequency of detected faces in a video, 
a potentially significant factor when analyzing videos in which human presence or 
facial expressions are relevant. Higher face detection rates may indicate a video's 
emphasis on human subjects or emotions, providing insight into the parties' emphasis 
on individual politicians or narratives centered on humans.
The 'Dominant Emotion Score' identifies the most prominent emotion expressed in the 
video. The dominant emotion score, on the other hand, provides insight into the party's 
communication strategy by revealing the emotion they wish to convey. This 
interaction contributes to a comprehensive understanding of the strategic positioning 
of the parties.
The Dominant Emotion Score identifies the dominant emotion in the video. It aids in 
comprehending the emotional tone of the parties' social media campaigns, thereby 
shedding light on their communication strategies.
With the emotional diversity and sentiment disparity features, we explore the 
emotional landscape of the video content in greater depth. Our fourth feature, 
'Emotional_Diversity_FE', represents the variety of emotions expressed in the video, 
whereas our fifth feature, 'Sentiment_Disparity_FE', captures the disparity in 
sentiment between the comments and the recognized speeches. We capture the 
emotional complexity of the content and its reception by comprehending the range of 
emotions expressed and the sentiment difference between the videos and the 
comments. It offers a unique perspective on how effectively the party conveys its 
messages and how the audience interprets them.
Diversity of Emotions: This is a measurement of the variety of emotions displayed in 
a video. Greater emotional diversity is indicative of a more complex emotional 
narrative. This could help determine whether a party is focusing on eliciting a single 
emotion or employing a more diverse emotional strategy.
22
This illustrates the disparity in tone between the comments and the recognized 
speeches in the videos. A positive value indicates that the comments are more positive 
than the speeches, while a negative value indicates the opposite. It enables the 
potential understanding of differences in sentiment perception between spoken content 
and textual comments.
Sixth on our list, the 'Engagement_Per_Second_FE' feature quantifies the average 
engagement per second. This feature can reveal whether the party's videos are 
engrossing enough to maintain the audience's attention throughout, or whether they 
lose their appeal over time.
Engagement Per Second: This metric determines the average engagement per second 
for a video. Higher values indicate higher levels of engagement or interaction during 
a given time period. This could be useful for comparing the engagement levels of 
videos with varying lengths or identifying segments with higher viewer engagement.
The 'Topic_Alignment_Score_FE' function measures the semantic alignment between 
the topics in the videos and the comments. This feature reveals whether the audience 
is in sync with the party's discussions or if there is a disconnect. It is a key indicator 
of the party's appeal to its audience.