Add files via upload
The objective of the final layer of machine learning analysis is to cluster videos based 
on the specific features extracted or produced in earlier layers. The primary purpose 
of clustering in this instance is to comprehend the strategic maneuvers employed by 
the analyzed political parties. The KMeans clustering algorithm, the Elbow method, 
and the StandardScaler method, all from the fields of machine learning and statistics, 
are the primary tools utilized in this layer to accomplish this objective.

This layer of analysis is of great importance, not only because it builds upon the robust 
and complex feature generation of previous layers, but also because it provides a 
concrete foundation for understanding the relative positions of the parties in terms of 
their videos and associated features. The critical insights that emerge from this layer 
include the identification of key video characteristics that the parties use strategically 
to define their positions or differentiate themselves. The analysis and resulting clusters 
facilitate the visualization of relationships and serve as a focal point for the 
identification of strategic moves.
This layer presents, in summary, a methodical strategy for discerning the strategic 
moves of political parties through the clustering of their videos based on their 
generated features. It employs sophisticated machine learning techniques, rigorous 
optimization procedures, and careful interpretation of results, all of which serve to 
improve our understanding of these parties' strategic positioning. In addition, the 
25
layer emphasizes the importance of ensuring equal feature contribution in the 
clustering process, identifies the optimal number of clusters, and paves the way for 
strategic insights to be gleaned from the videos of these political parties. This layer's 
methodology and findings provide a solid foundation for future research and have 
significant potential for application in political strategy analysis.